LayoutTests/platform/ios-simulator/fast/events/platform-wheelevent-paging-xy-in-scrolling-page-expected.txt

FAIL document.body.scrollTop should be 511. Was 0.
FAIL document.body.scrollLeft should be 686. Was 0.
